requirements.txt
agency-swarm==0.1.7;
python-dotenv;
openai;
instructor;
pydantic;